🐛 Fix code to work with 32-bits architecture

+++ src/predictor/cpu_predictor.cc
@@ -169,7 +169,7 @@ struct GHistIndexMatrixView {
   SparsePage::Inst operator[](size_t r) {
     auto t = omp_get_thread_num();
     auto const beg = (n_features_ * kUnroll * t) + (current_unroll_[t] * n_features_);
-    size_t non_missing{beg};
+    size_t non_missing{static_cast<size_t>(beg)};
 
     for (bst_feature_t c = 0; c < n_features_; ++c) {
       float f = page_.GetFvalue(ptrs_, values_, mins_, r, c, common::IsCat(ft_, c));
